pyproject.toml
[project]
name = "rag-production-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.135.1",
    "flask>=3.1.3",
    "google-genai>=1.67.0",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.17",
    "llama-index-readers-file>=0.6.0",
    "logging>=0.4.9.6",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
